# loyalty-platform-impl/openapi.yaml
openapi: "3.0.3"
info:
  title: Loyalty Banking Platform — In-Scope API (Capstone)
  description: |
    Public contract for the four I-11 use cases (UC-LB-01 through UC-LB-04).
    Scope: I-11 only. Names are Lab 1 I-4 container names.
    I-3 external systems (Core Banking System, Partner Systems, CRM & Notification Gateway,
    Enterprise Data Warehouse) are simulated at the integration boundary; no production credentials
    or real host names are used.

    **Constraint reference:**
    - CON.1 — No duplicate point posting for the same source transaction
    - CON.2 — No direct database writes from channels or other services
    - CON.3 — Fulfillment failure must compensate by restoring points (FIFO earn date and expiry preserved)
    - CON.4 — Analytics reporting data must not exceed 10 minutes staleness

    **Spec source:** lab3-spec.md §4 contract register (CT-04, CT-11, CT-12, CT-26)
    **RACI:** R Dev (Lê Huy Du) · A SA (Vũ Trường Quang) · C Test (Lê Huy Du)
  version: "1.0.0"
  contact:
    name: Team 2 — Vũ Trường Quang, Khuất Duy Bách, Đặng Duy Hoàng, Lê Huy Du
servers:
  - url: http://localhost:8081
    description: Earning Engine Service (I-4, Domain Services Zone)
  - url: http://localhost:8082
    description: Tiering System Service (I-4, Domain Services Zone)
  - url: http://localhost:8083
    description: Redemption Engine Service (I-4, Domain Services Zone)
  - url: http://localhost:8085
...
paths:
...
  /api/v1/partners/earn:
    post:
      operationId: recordEarn
      tags:
        - UC-LB-01 Earning Engine Service
      summary: UC-LB-01 happy path — record a settled earn event
      description: |
        Happy path: Core Banking System publishes a settled transaction event via API Gateway.
        Earning Engine Service validates the event, checks idempotency against Idempotency Store (CON.1),
        calculates base and campaign points, writes a PointTransaction to Earning DB, and publishes
        a QP accrual event to Message Broker.

        Named alts (lab3-spec.md §3 sequence and §5 exceptions):
        - **ALT / CON.1 duplicate (EXC-01)** — same sourceTxnId arrives again → 409 Conflict,
          no second PointTransaction written, original result returned (lab10 G6-A01).
        - **ALT / CON.1 reversal (EXC-02)** — Core Banking reverses a rewarded transaction →
          PointTransaction cancelled via PATCH /api/v1/partners/earn/{sourceTxnId}/cancel.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EarnEventRequest'
            example:
              transactionId: "txn-abc-123"
              memberId: "member-001"
              spendAmount: 1000
              tier: "SILVER"
              campaignId: "DOUBLE_POINTS_AUG"
              programId: "DEFAULT_PROG"
      responses:
        "202":
          description: |
            Accepted. PointTransaction written; QP accrual published to Message Broker.
            Happy path UC-LB-01.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EarnEventResponse'
        "409":
          description: |
            Conflict — CON.1 duplicate event detected (EXC-01, G6-A01).
            The same sourceTxnId has already been rewarded.
            No second PointTransaction is written; the original result is returned.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
              example:
                type: "https://loyalty.internal/errors/duplicate-transaction"
                title: "Duplicate Transaction"
                status: 409
                detail: "ERR_EARN_DUPLICATE: Transaction txn-abc-123 already rewarded under CON.1"
                instance: "/api/v1/partners/earn"
        "400":
          description: Bad request — missing or invalid fields.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /api/v1/partners/earn/{sourceTxnId}/cancel:
    patch:
      operationId: cancelEarnTransaction
      tags:
        - UC-LB-01 Earning Engine Service
      summary: UC-LB-01 alt — Core Banking reversal of a rewarded transaction (EXC-02, CON.1)
      description: |
        Named alt: Core Banking System reverses a source transaction that has already been rewarded.
        The matching PointTransaction is cancelled and PointBalance is corrected.
        Lab3 EXC-02 / CON.1.
      parameters:
        - in: path
          name: sourceTxnId
          required: true
          schema:
            type: string
          description: The original Core Banking source transaction ID to reverse.
      responses:
        "200":
          description: Reversal accepted; PointTransaction cancelled and PointBalance corrected.
          content:
            application/json:
              schema:
                type: object
                properties:
                  sourceTxnId:
                    type: string
                  outcome:
                    type: string
                    enum: [CANCELLED]
        "404":
          description: No matching PointTransaction found for the given sourceTxnId.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'

  # ── CT-13: DebitPointsFifo (Redemption Engine Service → Earning Engine Service)
  /api/v1/earning/fifo/debit:
    post:
      operationId: debitPointsFifo
      tags:
        - UC-LB-01 Earning Engine Service
      summary: CT-13 — Debit points in FIFO order from oldest unexpired batches
      description: |
        Contract CT-13: Redemption Engine Service delegates FIFO debit to Earning Engine Service.
        PointTransaction and PointBalance are owned by Earning Engine Service (I-7 single ownership).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FifoDebitRequest'
      responses:
        "200":
          description: FIFO debit reserved successfully in Earning DB.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FifoDebitResponse'
        "422":
          description: Insufficient balance in Earning DB.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        "503":
          description: Earning Engine Service unavailable.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'

  # ── CT-13: RestorePoints (CON.3 / EXC-05 Reversal)
  /api/v1/earning/fifo/restore:
    post:
      operationId: restorePoints
      tags:
        - UC-LB-01 Earning Engine Service
      summary: CT-13 — Auto-reversal restoring points to original batches with earn date preserved (CON.3 / EXC-05)
      description: |
        Contract CT-13 / EXC-05: When partner fulfillment fails, Redemption Engine Service
        calls Earning Engine Service to restore the debited points.
        Original earn_date and expiry_date are preserved under CON.3.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FifoRestoreRequest'
      responses:
        "200":
          description: Points restored successfully to original FIFO batches.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FifoRestoreResponse'
        "422":
          description: Unknown or invalid persisted debit allocation.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'

  # ── CT-13 Read Path: Get Member Balance
  /api/v1/earning/members/{memberId}/balance:
    get:
      operationId: getMemberBalance
      tags:
        - UC-LB-01 Earning Engine Service
      summary: CT-13 — Query confirmed point balance from Earning DB (I-7 Source of Truth)
      parameters:
        - in: path
          name: memberId
          required: true
          schema:
            type: string
        - in: query
          name: programId
          schema:
            type: string
            default: "DEFAULT_PROG"
      responses:
        "200":
          description: Current confirmed balance.
          content:
            application/json:
              schema:
                type: object
                properties:
                  memberId:
                    type: string
                  programId:
                    type: string
                  confirmedBalance:
                    type: integer

  # ─────────────────────────────────────────────────────────────────
  # UC-LB-02 — Redeem reward with FIFO (CT-11 CreateRedemptionOrder)
  # Server: http://localhost:8083
  # ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    EarnEventRequest:
      type: object
      required: [transactionId, memberId, spendAmount]
      properties:
        transactionId:
          type: string
          description: Idempotency key — source transaction ID from Core Banking System (CON.1).
          example: "txn-abc-123"
        memberId:
          type: string
          description: Lab 1 I-2 Member identifier (simulated).
          example: "member-001"
        spendAmount:
          type: integer
          description: Transaction spend amount in base currency units.
          example: 1000
        tier:
          type: string
          description: Current MemberTier at time of earning.
          enum: [SILVER, GOLD, PLATINUM]
          example: "SILVER"
        campaignId:
          type: string
          nullable: true
          description: Optional campaign ID for bonus point calculation.
          example: "DOUBLE_POINTS_AUG"
        programId:
          type: string
          description: Loyalty program identifier.
          example: "DEFAULT_PROG"
    EarnEventResponse:
      type: object
      properties:
        sourceTxnId:
          type: string
        memberId:
          type: string
        basePoints:
          type: integer
        bonusPoints:
          type: integer
        totalPoints:
          type: integer
        outcome:
          type: string
          enum: [CONFIRMED]

    # ── UC-LB-02 ─────────────────────────────────────────
...
    FifoDebitRequest:
      type: object
      required: [memberId, pointsRequired, orderId]
      properties:
        memberId:
          type: string
          example: "member-001"
        programId:
          type: string
          default: "DEFAULT_PROG"
        pointsRequired:
          type: integer
          minimum: 1
          example: 300
        orderId:
          type: string
          example: "order-123"
    FifoDebitResponse:
      type: object
      properties:
        orderId:
          type: string
        memberId:
          type: string
        pointsDebited:
          type: integer
        status:
          type: string
          enum: [RESERVED]
        debitedBatchesCount:
          type: integer
        allocationId:
          type: string
          description: Stable persisted allocation identity used by RestorePoints.
        message:
          type: string
    FifoRestoreRequest:
      type: object
      required: [memberId, orderId, pointsToRestore]
      properties:
        memberId:
          type: string
        programId:
          type: string
          default: "DEFAULT_PROG"
        orderId:
          type: string
        allocationId:
          type: string
        pointsToRestore:
          type: integer
          minimum: 1
        reason:
          type: string
          example: "OUT_OF_STOCK"
    FifoRestoreResponse:
      type: object
      properties:
        orderId:
          type: string
        memberId:
          type: string
        pointsRestored:
          type: integer
        status:
          type: string
          enum: [RESTORED]
        reason:
          type: string
        allocationId:
          type: string
          description: Stable persisted allocation identity used for idempotent restoration.
        message:
          type: string
...
    ProblemDetail:
      type: object
      description: RFC 7807 problem detail.
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string